Move just-next-tick back to iter/
